+++ src/Modules/glue_meddly.h
@@ -38,7 +38,7 @@ class smart_output : public MEDDLY::output {
     virtual void put(unsigned long x, int w);
     virtual void put_hex(unsigned long x, int w);
     virtual void put(double x, int w, int p, char f);
-    virtual int write(int bytes, const unsigned char* buffer);
+    virtual size_t write(size_t bytes, const unsigned char* buffer);
     virtual void flush();
 
   private:
